+++ lib/cfn.js
@@ -3,6 +3,7 @@ const path = require('path');
 const webhooks = require('./events/webhook');
 const snsEvent = require('./events/sns');
 const cloudWatchEvent = require('./events/cloudwatch');
+const roles = require('./roles');
 
 module.exports = buildFunctionTemplate; // returns full template
 module.exports.embed = buildFunction; // returns only lambda resources for template merge
@@ -177,7 +178,7 @@ function buildFunction(options) {
   var parameters = buildParameters(options);
   var defaultAlarms = buildServiceAlarms(options);
   var lambda = buildLambda(options);
-  var role = buildRole(options);
+  var role = roles.buildRole(options);
   var cloudwatchEvent = {};
   var schedule = {};
   var sns = {};
@@ -388,151 +389,6 @@ function buildLambda(options) {
   return fn;
 }
 
-module.exports.buildRole = buildRole;
-function buildRole(options) {
-  var role = {
-    Resources: {}
-  };
-
-  role.Resources['LambdaCfnRole'] = {
-    Type: 'AWS::IAM::Role',
-    Properties: {
-      AssumeRolePolicyDocument: {
-        Statement: [
-          {
-            Sid: '',
-            Effect: 'Allow',
-            Principal: {
-              Service: 'lambda.amazonaws.com'
-            },
-            Action: 'sts:AssumeRole'
-          },
-          {
-            Sid: '',
-            Effect: 'Allow',
-            Principal: {
-              Service: 'events.amazonaws.com'
-            },
-            Action: 'sts:AssumeRole'
-          }
-        ]
-      },
-      Path: '/',
-      Policies: [
-        {
-          PolicyName: 'basic',
-          PolicyDocument: {
-            Statement: [
-              {
-                Effect: 'Allow',
-                Action: [
-                  'logs:*'
-                ],
-                Resource: cf.join(['arn:aws:logs:*:', cf.accountId, ':*'])
-              },
-              {
-                Effect: 'Allow',
-                Action: [
-                  'sns:Publish'
-                ],
-                Resource: cf.ref('ServiceAlarmSNSTopic')
-              },
-              {
-                Effect: 'Allow',
-                Action: [
-                  'iam:SimulateCustomPolicy'
-                ],
-                Resource: '*'
-              }
-            ]
-          }
-        },
-      ]
-    }
-  };
-
-  role.Resources['DispatchRole'] = {
-    Type: 'AWS::IAM::Role',
-    Condition: 'HasDispatchSnsArn',
-    Properties: {
-      AssumeRolePolicyDocument: {
-        Statement: [
-          {
-            Sid: '',
-            Effect: 'Allow',
-            Principal: {
-              Service: 'lambda.amazonaws.com'
-            },
-            Action: 'sts:AssumeRole'
-          },
-          {
-            Sid: '',
-            Effect: 'Allow',
-            Principal: {
-              Service: 'events.amazonaws.com'
-            },
-            Action: 'sts:AssumeRole'
-          }
-        ]
-      },
-      Path: '/',
-      Policies: [
-        {
-          PolicyName: 'basic',
-          PolicyDocument: {
-            Statement: [
-              {
-                Effect: 'Allow',
-                Action: [
-                  'sns:Publish'
-                ],
-                Resource: cf.ref('DispatchSnsArn')
-              },
-            ]
-          }
-        },
-      ]
-    }
-  };
-
-  if (options.eventSources && options.eventSources.webhook) {
-    role.Resources['LambdaCfnRole'].Properties.AssumeRolePolicyDocument.Statement.push({
-      Sid: '',
-      Effect: 'Allow',
-      Principal: {
-        Service: 'apigateway.amazonaws.com'
-      },
-      Action: 'sts:AssumeRole'
-    });
-  }
-
-  if (options.statements) {
-    let areValidStatementsOptions = !Array.isArray(options.statements);
-    if (areValidStatementsOptions) {
-      throw new Error('options.statements must be an array');
-    }
-  }
-
-  if (options.statements) {
-    // Very basic validation on each policy statement
-    options.statements.forEach(function(statement) {
-      if (!statement.Effect)
-        throw new Error('statement must contain Effect');
-      if (!statement.Resource && !statement.NotResource)
-        throw new Error('statement must contain Resource or NotResource');
-      if (!statement.Action && !statement.NotAction)
-        throw new Error('statement must contain Action or NotAction');
-    });
-    role.Resources['LambdaCfnRole'].Properties.Policies.push({
-      PolicyName: options.name,
-      PolicyDocument: {
-        Statement: options.statements
-      }
-    });
-  }
-
-  return role;
-}
 
 module.exports.buildServiceAlarms = buildServiceAlarms;
 function buildServiceAlarms(options) {
@@ -656,3 +512,4 @@ function buildSnsDestination(options) {
 module.exports.buildWebhookEvent = webhooks.buildWebhookEvent;
 module.exports.buildSnsEvent = webhooks.buildSnsEvent;
 module.exports.buildCloudwatchEvent = cloudWatchEvent.buildCloudwatchEvent;
+module.exports.buildRole = roles.buildRole;
